Added code action filter/autorun method and cs bindings.
I repeatedly run code actions, then search for "Generate field ...".
This is tiring and pointless as I almost always need to search for a
specific option, and the ordering changes. Wrote a core method to pass a
filter string into code action, which will auto run if there is only one
option.

Made a keybinding for csharp to use it as well.